Introduce parent class to just have to override the important parts like articles, header, summary
